+++ include/parameters/FrequencyParameter.h
@@ -3,7 +3,7 @@
 
 #include "midi_helpers.h"
 
-template<class TargetClass, class DataType = double>
+template<class TargetClass, class DataType = float>
 class FrequencyParameter : public Parameter<TargetClass, DataType> {
     public:
 
@@ -14,7 +14,7 @@ class FrequencyParameter : public Parameter<TargetClass, DataType> {
             {
             this->octave_range = octave_range;
         };
-        FrequencyParameter(char *label, TargetClass *target, double initial_value, void(TargetClass::*setter_func)(DataType), DataType octave_range = 5.0) 
+        FrequencyParameter(char *label, TargetClass *target, float initial_value, void(TargetClass::*setter_func)(DataType), DataType octave_range = 5.0) 
             : Parameter<TargetClass,DataType>(label, target, initial_value, setter_func) {
                 this->octave_range = octave_range;
             };
@@ -34,7 +34,7 @@ class FrequencyParameter : public Parameter<TargetClass, DataType> {
 
             this->octave_range = octave_range;
 
-            double freq = get_frequency_for_voltage(value * this->octave_range); //read_voltage(0));
+            float freq = get_frequency_for_voltage(value * this->octave_range); //read_voltage(0));
 
             this->lastValue = this->currentValue;
             this->currentValue = value;
@@ -55,9 +55,9 @@ class FrequencyParameter : public Parameter<TargetClass, DataType> {
         virtual const char* getFormattedValue() {
             static char fmt[20] = "              ";
 
-            double voltage = this->currentValue * this->octave_range;
+            float voltage = this->currentValue * this->octave_range;
 
-            double freq = get_frequency_for_voltage(voltage); //read_voltage(0));
+            float freq = get_frequency_for_voltage(voltage); //read_voltage(0));
             int pitch   = get_midi_pitch_for_voltage(voltage);
 
             //Serial.printf("%s: getFormattedValue got pitch %i: %s\n", this->label, pitch, get_note_namec(pitch));
