+++ include/voltage_sources/ADS24vVoltageSource.h
@@ -23,13 +23,13 @@ class ADS24vVoltageSource : public ADSVoltageSource<ADS1X15Type> {
                 //          1181, 0.0400
         }
 
-        virtual double adcread_to_voltage(int16_t adcReading) override {
+        virtual float adcread_to_voltage(int16_t adcReading) override {
             // for some reason, seem to have to do this calculation differently (than the parent class) for this board?
             //return ads_source->toVoltage(adcReading);
-            double voltageCorrected = ((double)adcReading/this->correction_value_1) - 1.0;
+            float voltageCorrected = ((float)adcReading/this->correction_value_1) - 1.0;
             /*
             // tryna implement same logic as https://github.com/pimoroni/ads1015-python/blob/9ae775bc0a4a148771ef1d36a98adb76cc6cf54e/library/ads1015/__init__.py#L375 
-            double voltageCorrected = adcReading * 1000.0;
+            float voltageCorrected = adcReading * 1000.0;
             voltageCorrected /= 2048.0;
             voltageCorrected *= this->ads_source->getGain();
             voltageCorrected /= 1000.0;
@@ -42,14 +42,14 @@ class ADS24vVoltageSource : public ADSVoltageSource<ADS1X15Type> {
             return voltageCorrected;
         }
 
-        virtual double get_corrected_voltage(double voltageFromAdc) override {
-            double corrected2 = voltageFromAdc / this->correction_value_2;
+        virtual float get_corrected_voltage(float voltageFromAdc) override {
+            float corrected2 = voltageFromAdc / this->correction_value_2;
             /*
             // tryna implement same logic as https://github.com/pimoroni/ads1015-python/blob/9ae775bc0a4a148771ef1d36a98adb76cc6cf54e/library/ads1015/__init__.py#L375 
-            double pin_v = voltageFromAdc;
-            double input_v = pin_v * ((float)8060000+402000) / (float)8060000;
+            float pin_v = voltageFromAdc;
+            float input_v = pin_v * ((float)8060000+402000) / (float)8060000;
             input_v += 1.241;
-            double corrected2 = input_v;
+            float corrected2 = input_v;
             */
             /*if (this->debug) {
                 Serial.printf("after applying correction stage 1, got ");
